djangoProject/requirements.txt
Django==5.1.3
mysql-connector-python==9.1.0
mysqlclient==2.1.1
requests==2.25.1
pytz==2022.1
sqlparse==0.5.2
typing_extensions==4.12.2